feat(fp): 优雅支持 rv64ud 双精度浮点指令集（inorder/ooo）
本次提交在保持 rv64uf 行为稳定的前提下，完整打通 rv64ud-p-* 测试路径。

核心改动：

- 将 inorder/ooo 的浮点寄存器统一为 64 位，符合 FLEN=64 语义。

- 在浮点访存路径引入 NaN-boxing：FLW 写入高 32 位全 1，FSW 仅写低 32 位。

- 扩展 InstructionExecutor 浮点执行器为 fmt(S/D) 分流，补齐 D 扩展运算：

  fadd/fsub/fmul/fdiv/fsqrt、fmin/fmax、fsgnj*、feq/flt/fle、

  fcvt（含 S<->D 与 int<->S/D）、fmv/fclass、fma 家族（FMADD/FMSUB/FNMSUB/FNMADD）。

- 调整 FPU 执行接口入参到 64 位位模式，统一 S/D 执行路径。

- 在 Decoder 中补充 D 扩展识别与校验，避免仅通过 F 扩展放行 D 指令。

- 默认启用扩展增加 D（inorder + ooo + CPUState 默认配置）。

- 更新相关单测断言以匹配 NaN-boxed 单精度结果。

验证结果：

- cmake -S . -B build && cmake --build build -j8

- ctest --test-dir build --output-on-failure

- python3 run_tests.py -p "rv64ud-p-*"

- python3 run_tests.py -p "rv64ud-p-*" --ooo

- python3 run_tests.py -p "rv64uf-p-*"

- python3 run_tests.py -p "rv64uf-p-*" --ooo

+++ include/core/instruction_executor.h
@@ -59,11 +59,11 @@ public:
     
     // F扩展指令（单精度浮点）
     static uint32_t executeFPExtension(const DecodedInstruction& inst, float rs1_val, float rs2_val);
-    static FpExecuteResult executeFPOperation(const DecodedInstruction& inst, uint32_t rs1_bits,
-                                              uint32_t rs2_bits, uint64_t rs1_int,
+    static FpExecuteResult executeFPOperation(const DecodedInstruction& inst, uint64_t rs1_bits,
+                                              uint64_t rs2_bits, uint64_t rs1_int,
                                               uint8_t current_frm = 0);
-    static FpExecuteResult executeFusedFPOperation(const DecodedInstruction& inst, uint32_t rs1_bits,
-                                                   uint32_t rs2_bits, uint32_t rs3_bits,
+    static FpExecuteResult executeFusedFPOperation(const DecodedInstruction& inst, uint64_t rs1_bits,
+                                                   uint64_t rs2_bits, uint64_t rs3_bits,
                                                    uint8_t current_frm = 0);
     static bool isFPIntegerDestination(const DecodedInstruction& inst);
     static bool isFloatingPointInstruction(const DecodedInstruction& inst);
